+++ app/upgrades/btcstaking/delete_dup_fp_test.go
@@ -1,4 +1,4 @@
-package upgrades
+package btcstaking_test
 
 import (
 	"context"
@@ -11,8 +11,15 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 
+<<<<<<< HEAD:app/upgrades/delete_dup_fp_test.go
 	bbn "github.com/babylonlabs-io/babylon/v3/types"
 	btcstktypes "github.com/babylonlabs-io/babylon/v3/x/btcstaking/types"
+=======
+	_ "github.com/babylonlabs-io/babylon/v4/app/params" // import this to run the init func that sets the address prefixes
+	"github.com/babylonlabs-io/babylon/v4/app/upgrades/btcstaking"
+	bbn "github.com/babylonlabs-io/babylon/v4/types"
+	btcstktypes "github.com/babylonlabs-io/babylon/v4/x/btcstaking/types"
+>>>>>>> d79f7c56 (imp(btcstkconsumer): add finality contract idx (#1596)):app/upgrades/btcstaking/delete_dup_fp_test.go
 )
 
 type FpFromTestData struct {
@@ -116,7 +123,7 @@ func TestFpSoftDeleteDupAddr(t *testing.T) {
 			}
 
 			ctx := context.Background()
-			err = FpSoftDeleteDupAddr(ctx, mockBtcStkKeeper)
+			err = btcstaking.FpSoftDeleteDupAddr(ctx, mockBtcStkKeeper)
 			require.NoError(t, err)
 			// Verify that SetFpBbnAddr was called for each FP
 			require.EqualValues(t, mockBtcStkKeeper.setBbnAddrsCount, len(fps), "SetFpBbnAddr should be called for each FP")