+++ app/upgrades/v2_3/upgrades.go
@@ -7,8 +7,14 @@ import (
 	upgradetypes "cosmossdk.io/x/upgrade/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 
+<<<<<<< HEAD
 	"github.com/babylonlabs-io/babylon/v3/app/keepers"
 	"github.com/babylonlabs-io/babylon/v3/app/upgrades"
+=======
+	"github.com/babylonlabs-io/babylon/v4/app/keepers"
+	"github.com/babylonlabs-io/babylon/v4/app/upgrades"
+	"github.com/babylonlabs-io/babylon/v4/app/upgrades/btcstaking"
+>>>>>>> d79f7c56 (imp(btcstkconsumer): add finality contract idx (#1596))
 )
 
 // UpgradeName defines the on-chain upgrade name for the Babylon v2.3 upgrade
@@ -30,7 +36,7 @@ func CreateUpgradeHandler(mm *module.Manager, configurator module.Configurator, 
 			return nil, err
 		}
 
-		err = upgrades.FpSoftDeleteDupAddr(ctx, keepers.BTCStakingKeeper)
+		err = btcstaking.FpSoftDeleteDupAddr(ctx, keepers.BTCStakingKeeper)
 		if err != nil {
 			return nil, err
 		}